refactor: decompose bending.py and fix regressions in selection helpers
